requirements.txt
langchain==0.3.23
langchain-openai==0.3.14
python-dotenv==1.1.0
